requirements.txt
librosa
soundfile
numpy
torch
pyannote.audio
python-dotenv
transformers
accelerate
